Inline Chat: Separate from sidebar (and ChatProvider) (#7)
