+++ vscode/CHANGELOG.md
@@ -9,6 +9,9 @@ Starting from `0.2.0`, Cody is using `major.EVEN_NUMBER.patch` for release versi
 ### Added
 
 - [Internal] `Custom Recipes`: An experimental feature now available behind the `cody.experimental.customRecipes` feature flag for internal testing purpose. [pull/348](https://github.com/sourcegraph/cody/pull/348)
+- Inline Chat: Improved response quality by ensuring each inline chat maintains its own unique context, and doesn't share with the sidebar and other inline chats. This should also benefit response quality for inline /fix and /touch commands.
+- Inline Chat: Added the option to 'Stop generating' from within the inline chat window.
+- Inline Chat: Added the option to transfer a chat from the inline window to the Cody sidebar.
 
 ### Fixed
